format table

+++ bamazon_db.sql
@@ -6,16 +6,16 @@ CREATE TABLE products (
     item_id INTEGER(10) NOT NULL AUTO_INCREMENT,
     product_name VARCHAR(200) NOT NULL,
     department_name VARCHAR(50) DEFAULT "other",
-    price DECIMAL(10,2) NOT NULL,
+    price DOUBLE(10,2) NOT NULL,
     stock_quantity INTEGER(10) NOT NULL,
     sold_quantity INTEGER(10) DEFAULT 0,
-    product_sales DECIMAL(10,2) DEFAULT 0,
+    product_sales DOUBLE(10,2) DEFAULT 0,
     PRIMARY KEY (item_id)
 );
 CREATE TABLE departments (
     department_id INTEGER(10) NOT NULL AUTO_INCREMENT,
     department_name VARCHAR(200) NOT NULL,
-    over_head_costs DECIMAL(10,2) NOT NULL,
+    over_head_costs DOUBLE(10,2) NOT NULL,
     PRIMARY KEY (department_id)
 );
